requirements.txt
flask
sqlalchemy
psycopg2
spacy
python-dotenv
pdfminer.six
